requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
bcrypt
pydantic
pandas
openpyxl
psycopg2-binary
gunicorn
